fix: fixed issue with filters being disabled on selecting a component
Localized filters provider to the header of the asset view so that it
doesn't cause unnecessary re-renders on the asset tree. This fixed the
issue where the default classes were restored on re-render, causing the
filters to be disabled. Moving the elements out of the asset view
required some style adjustments as well.